# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AppStore API",
    "version": "2.0.0",
    "description": "공개형 Developer App Store의 REST API. 브라우저 세션은 CSRF 헤더를 사용하고 자동화 클라이언트는 권한이 제한된 개인 API Key를 사용합니다."
  },
  "servers": [{"url": "/", "description": "Same-origin AppStore server"}],
...
  "paths": {
...
    "/api/v1/apps": {
      "get": {
        "tags":["Public"],"summary":"공개 앱 목록/검색","operationId":"listApps",
        "parameters":[
          {"name":"q","in":"query","schema":{"type":"string"}},
          {"name":"category","in":"query","schema":{"type":"string"}},
          {"name":"language","in":"query","schema":{"type":"string"}},
          {"name":"mcp","in":"query","schema":{"type":"boolean"}},
          {"name":"featured","in":"query","schema":{"type":"boolean"}},
          {"name":"sort","in":"query","schema":{"type":"string","enum":["updated","trending","name","created"]}},
          {"name":"limit","in":"query","schema":{"type":"integer","minimum":1,"maximum":100,"default":24}},
          {"name":"offset","in":"query","schema":{"type":"integer","minimum":0,"default":0}}
        ],
        "responses":{"200":{"description":"Application page","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AppPage"}}}},"400":{"$ref":"#/components/responses/Error"}}
      },
      "post": {
        "tags":["Personal"],"summary":"앱 등록","operationId":"createApp","security":[{"cookieSession":[]},{"apiKey":[]}],
        "requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AppInput"}}}},
        "responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/App"}}}},"401":{"$ref":"#/components/responses/Error"},"422":{"$ref":"#/components/responses/Error"}}
      }
    },
...
    "/api/v1/admin/users": {
      "get": {"tags":["Admin"],"summary":"사용자 목록","operationId":"listUsers","security":[{"cookieSession":[]}],"responses":{"200":{"description":"Users"}}}
    },
    "/api/v1/admin/users/{id}": {
      "put": {"tags":["Admin"],"summary":"사용자 정보·활성 상태 변경","operationId":"updateUser","security":[{"cookieSession":[]}],"parameters":[{"$ref":"#/components/parameters/ID"}],"responses":{"200":{"description":"Updated user"}}},
      "delete": {"tags":["Admin"],"summary":"감사 이력을 보존하며 사용자 접근 제거","operationId":"deleteUser","security":[{"cookieSession":[]}],"parameters":[{"$ref":"#/components/parameters/ID"}],"responses":{"204":{"description":"Access removed"}}}
    },
...
    "/api/v1/admin/mcp": {
      "get": {"tags":["Admin"],"summary":"MCP 2026-07-28 정책","operationId":"getMCPSettings","security":[{"cookieSession":[]}],"responses":{"200":{"description":"MCP settings"}}},
      "put": {"tags":["Admin"],"summary":"MCP 정책 변경","operationId":"updateMCPSettings","security":[{"cookieSession":[]}],"responses":{"200":{"description":"Updated"}}}
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieSession": {"type":"apiKey","in":"cookie","name":"appstore_session"},
      "apiKey": {"type":"http","scheme":"bearer","bearerFormat":"aps_..."}
    },
    "parameters": {
      "ID": {"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},
...
    },
    "responses": {
      "Error": {"description":"Standard error envelope","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}
    },
    "schemas": {
...
      "Category": {"type":"object","required":["id","slug","name","icon","appCount"],"properties":{"id":{"type":"string","format":"uuid"},"slug":{"type":"string"},"name":{"type":"string"},"icon":{"type":"string"},"description":{"type":"string"},"position":{"type":"integer"},"appCount":{"type":"integer"}}},
...
      "App": {"type":"object","required":["id","name","slug","summary","description","icon","categoryId","tags","screenshots","supportsMcp","supportsApi","visibility","status"],"properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"slug":{"type":"string"},"summary":{"type":"string"},"description":{"type":"string"},"icon":{"type":"string"},"gradient":{"type":"string"},"serviceUrl":{"type":"string","format":"uri"},"category":{"$ref":"#/components/schemas/Category"},"categoryId":{"type":"string","format":"uuid"},"tags":{"type":"array","items":{"type":"string"}},"screenshots":{"type":"array","items":{"type":"string"}},"language":{"type":"string"},"framework":{"type":"string"},"supportsMcp":{"type":"boolean"},"supportsApi":{"type":"boolean"},"ownerId":{"type":"string","format":"uuid"},"team":{"type":"string"},"version":{"type":"string"},"visibility":{"type":"string","enum":["public","private"]},"status":{"type":"string","enum":["draft","pending_review","published","rejected","archived"]},"featured":{"type":"boolean"},"trendingScore":{"type":"integer"}}},
      "AppInput": {"type":"object","required":["name","slug","summary","description","icon","serviceUrl","categoryId","visibility"],"properties":{"name":{"type":"string","minLength":2,"maxLength":120},"slug":{"type":"string","pattern":"^[a-z0-9]+(?:-[a-z0-9]+)*$"},"summary":{"type":"string","maxLength":240},"description":{"type":"string","maxLength":20000},"icon":{"type":"string"},"gradient":{"type":"string"},"serviceUrl":{"type":"string","format":"uri"},"categoryId":{"type":"string","format":"uuid"},"tags":{"type":"array","maxItems":30,"items":{"type":"string"}},"screenshots":{"type":"array","maxItems":12,"items":{"type":"string"}},"language":{"type":"string"},"framework":{"type":"string"},"supportsMcp":{"type":"boolean"},"supportsApi":{"type":"boolean"},"team":{"type":"string"},"version":{"type":"string"},"visibility":{"type":"string","enum":["public","private"]}}},
      "AppPage": {"type":"object","required":["items","total","limit","offset"],"properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/App"}},"total":{"type":"integer"},"limit":{"type":"integer"},"offset":{"type":"integer"}}},
...
      "Error": {"type":"object","required":["error"],"properties":{"error":{"type":"object","required":["code","message","requestId"],"properties":{"code":{"type":"string"},"message":{"type":"string"},"requestId":{"type":"string"},"details":{"type":"object","additionalProperties":true}}}}}
    }
  }
}